RDKB-33035 - Fix warnings in Telemetry
Reason for change: To fix below warnings in Telemetry

1. implicit declaration of function <function name> [-Werror=implicit-function-declaration]
2. ignoring return value of 'realloc', declared with attribute warn_unused_result [-Wunused-result]
3. assignment makes pointer from integer without a cast [-Werror]
4. incompatible implicit declaration of built-in function <function name> [-Werror]
5. passing argument <argument number> of <function name> from incompatible pointer type [-Werror]
6. assignment discards 'const' qualifier from pointer target type [-Werror]
7. initialization discards 'const' qualifier from pointer target type [-Werror]
8. missing initializer for field <field name> of <struct name> [-Wmissing-field-initializers]
9. suggest parentheses around assignment used as truth value [-Wparentheses]
10. missing braces around initializer [-Wmissing-braces]
11. conflicting types for <function name> [-Werror]
12. backslash and newline separated by space [-Werror]

Test Procedure: Build should be successful without any warnings / errors.
Basic functionalities of Telemetry should be verified and its regression should pass.
Risks: Low

(cherry picked from commit 51f165539a64ae5dac763e54200dd853764e3b94)

+++ source/t2dm/cosa_telemetry_dml.c
@@ -140,7 +140,7 @@ BOOL Telemetry_SetParamStringValue(ANSC_HANDLE hInsContext, char* ParamName, cha
     if(AnscEqualString(ParamName, "ReportProfilesMsgPack", TRUE))
     {
         char *webConfigString  = NULL;
-        int stringSize = 0;
+        ULONG stringSize = 0;
 	if(pString != NULL)
 	{
  		webConfigString = AnscBase64Decode(pString, &stringSize);
